wrap: wrap the essl compiler.
This introduces a header with the api to
__mali_compile_essl_shader

And we now wrap this symbol to allow for printing of both the shader
source and the compiled (but not yet linked) shader binary and
linker streams. Since we have the actual GP/PP memory dumps as well,
we will be able to implement the linking ourselves (and complete it
over time).

Signed-off-by: Luc Verhaegen <libv@codethink.co.uk>