wrapper: fix recursive lock.
Apparently the kernel is not returning regularly on a newer
telechips image, it instead waits for a job to come in. But because
our dump needs to have things serialized, before sending oof the job,
the waiting had to return...

Also fix the rather poor implementation fbdev ioctl numbers which
made our applications fail to write anything to fbdev when using
the wrapper.

Signed-off-by: Luc Verhaegen <libv@codethink.co.uk>