premali: add smoothed strip test.
To get this strip working, we had to fix our symbol implementation
and the way the layout of those symbols got specificied in the vs
stream.
We now also figured out how array count, and draw mode are fed into
the vs and plbu.

Now we go and kick in the fan open door, just to get the test number
ramped up.

Signed-off-by: Luc Verhaegen <libv@codethink.co.uk>