limare: introduce varying maps
This the first step into properly understanding how varyings are passed
between the GP and the PP. By mapping things properly, we can now
correctly determine the flags and types that we need to pass into the
plbu. We treat gl_Position separately, as the PP treats this separately
and the plbu needs to have a special type for this.

Signed-off-by: Luc Verhaegen <libv@skynet.be>