limare: properly split vertex count between vs and plb
The VS simply works on all attributes, and outputs as many varyings and
vertices as there are attributes. Completely oblivious of the provided
elements list.

The PLBU then takes the elements list, or just _all_ or a given range of,
and drags the shaded attributes, or varyings, into the plbs.

We therefor have to be careful what we tell the VS to do, and about how
much space we really need for our varyings.

Signed-off-by: Luc Verhaegen <libv@skynet.be>