limare: fix bug in varying symbols creation
Initially, we had varying ordering dealt with differently,
and we therefor assigned the order in the table from the
offset provided by the symbol itself. This now crashed horribly
on a shader which was assigning this offset at a 2 aligned value,
which then broke with use using that offset, divided by four,
as the index to the table.

We now have proper varying mapping, which takes this offset into
account. The index into the symbol table is therefor irrelevant.

Signed-off-by: Luc Verhaegen <libv@skynet.be>