limare: add support for multiple textures
Optimization potential:
Texture units per draw are passed as uniforms, but internally they are
handled separately, as samplers. The symbol offset is not used as an
index into the uniform table, but instead used as an index in the
texture descriptor list for this draw. Space is still reserved in the
uniform table, for no good reason. Our own compiler could do better
and save some space.

Signed-off-by: Luc Verhaegen <libv@skynet.be>