limare: fix attribute count when not using indexed draws
When doing indexed draws, the GP transforms all attributes, and the plbu
then creates the plbs with those attributes which are indexed. Start and
count of the draw therefor describe the indexes, and not the actual
attributes. Those actual attributes are the size of the pointer or the
buffer, and are treated differently.

When doing direct draws, the start and count depict the actual entries in
the attribute, and the total attributes count should not be taken into
account. And it is exactly this bug that we now ran into.

Todo: copy only those attributes that are being used, for both the
direct and indexed draws. This makes the current bugfix clearer, wastes
less resources, and can be useful for both direct and indexed draws.

Signed-off-by: Luc Verhaegen <libv@skynet.be>